Added list sites feature
Completed the List Current Websites project, includes addition of the print_sites() function and the print_sites_verbose() function, as well as corresponding arguments for argparse.

The standard function prints a title for the data and the sites, whereas the verbose version formats it slightly, as well as listing how many sites are stored.